backend/requirements.txt
# FastAPI y servidor
fastapi
uvicorn
python-multipart

# Modelo de IA y procesamiento
torch
diffusers
transformers
Pillow
safetensors
imageio
imageio-ffmpeg

# Dependencias adicionales para el modelo Zeroscope
einops
omegaconf
